# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BirrStream API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  fullName: string;
  username: string;
  email: string;
  mainBalance: number;
  totalYield: number;
  totalDeposited: number;
  totalWithdrawn: number;
  referralCode: string;
  isAdmin: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface RegisterInput {
  fullName: string;
  username: string;
  email: string;
  password: string;
  confirmPassword: string;
  /** @nullable */
  referralCode?: string | null;
}

export interface LoginInput {
  usernameOrEmail: string;
  password: string;
  rememberMe?: boolean;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface DashboardSummary {
  mainBalance: number;
  totalYield: number;
  totalDeposited: number;
  totalWithdrawn: number;
  /** @nullable */
  activePackageName?: string | null;
  /** @nullable */
  activePackageDailyReturn?: number | null;
  /** @nullable */
  packageExpiresAt?: string | null;
  /** @nullable */
  daysUntilExpiry?: number | null;
  progressToNextTier: number;
  /** @nullable */
  nextTierName?: string | null;
  reserveFloor: number;
}

export interface StreakDay {
  dayNumber: number;
  date: string;
  checkedIn: boolean;
  bonus: number;
}

export interface LoginStreak {
  currentStreak: number;
  days: StreakDay[];
  todayCheckedIn: boolean;
  bonusPerDay: number;
}

export interface StreakCheckinResult {
  success: boolean;
  bonusEarned: number;
  newStreak: number;
  newBalance: number;
}

export type PackageTier = typeof PackageTier[keyof typeof PackageTier];


export const PackageTier = {
  vip1: 'vip1',
  vip2: 'vip2',
  vip3: 'vip3',
  vip4: 'vip4',
  vip5: 'vip5',
  elite: 'elite',
  apex: 'apex',
  titan: 'titan',
  alpha: 'alpha',
} as const;

export interface Package {
  id: number;
  name: string;
  cost: number;
  dailyReturn: number;
  totalYield: number;
  durationDays: number;
  isLocked: boolean;
  tier: PackageTier;
}

export interface PurchaseResult {
  success: boolean;
  package?: Package;
  newBalance?: number;
  message: string;
  /** @nullable */
  shortfallAmount?: number | null;
}

export interface ActivePackage {
  package: Package;
  purchasedAt: string;
  expiresAt: string;
  daysRemaining: number;
  totalEarnedSoFar: number;
  dailyEarningsToday: number;
}

export type DailyTaskTaskType = typeof DailyTaskTaskType[keyof typeof DailyTaskTaskType];


export const DailyTaskTaskType = {
  stream_video: 'stream_video',
  open_page: 'open_page',
  join_telegram: 'join_telegram',
  other: 'other',
} as const;

export interface DailyTask {
  id: number;
  title: string;
  description: string;
  reward: number;
  taskType: DailyTaskTaskType;
  /** @nullable */
  actionUrl?: string | null;
  /** @nullable */
  completedAt?: string | null;
  isCompleted: boolean;
}

export interface TaskCompleteResult {
  success: boolean;
  rewardEarned: number;
  newBalance: number;
  message: string;
}

export type DepositStatus = typeof DepositStatus[keyof typeof DepositStatus];


export const DepositStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Deposit {
  id: number;
  amount: number;
  senderName: string;
  /** @nullable */
  receiptUrl?: string | null;
  status: DepositStatus;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalDeposited: number;
  totalWithdrawn: number;
  pendingDepositsCount: number;
  pendingDepositsAmount: number;
  pendingWithdrawalsCount: number;
  pendingWithdrawalsAmount: number;
  totalActivePackages: number;
}

export type AdminDepositStatus = typeof AdminDepositStatus[keyof typeof AdminDepositStatus];


export const AdminDepositStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface AdminDeposit {
  id: number;
  userId: number;
  username: string;
  fullName: string;
  amount: number;
  senderName: string;
  /** @nullable */
  receiptUrl?: string | null;
  status: AdminDepositStatus;
  createdAt: string;
}

export type AdminWithdrawalStatus = typeof AdminWithdrawalStatus[keyof typeof AdminWithdrawalStatus];


export const AdminWithdrawalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface AdminWithdrawal {
  id: number;
  userId: number;
  username: string;
  fullName: string;
  amount: number;
  bankName: string;
  accountName: string;
  walletId: string;
  status: AdminWithdrawalStatus;
  createdAt: string;
}

export interface DepositInput {
  amount: number;
  senderName: string;
  /** @nullable */
  receiptBase64?: string | null;
}

export type WithdrawalStatus = typeof WithdrawalStatus[keyof typeof WithdrawalStatus];


export const WithdrawalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Withdrawal {
  id: number;
  amount: number;
  status: WithdrawalStatus;
  bankName: string;
  accountName: string;
  walletId: string;
  createdAt: string;
}

export interface WithdrawalInput {
  amount: number;
}

export interface WithdrawalSettings {
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  walletId?: string | null;
  isConfigured: boolean;
}

export interface WithdrawalSettingsInput {
  bankName: string;
  accountName: string;
  walletId: string;
}

export interface ReferralInfo {
  referralCode: string;
  referralLink: string;
  totalDirectReferrals: number;
  totalNetworkSize: number;
  totalCommissionsEarned: number;
  level1Count: number;
  level2Count: number;
  level3Count: number;
}

export interface AffiliateEntry {
  id: number;
  username: string;
  joinedAt: string;
  activeDepositAmount: number;
  commissionPaid: number;
  hasActivePackage: boolean;
}

export interface AffiliateNetwork {
  level1: AffiliateEntry[];
  level2: AffiliateEntry[];
  level3: AffiliateEntry[];
}

export interface VipUpgradeGoal {
  id: number;
  packageName: string;
  requiredDirectReferrals: number;
  requiredDownlineVolume: number;
  currentDirectReferrals: number;
  currentDownlineVolume: number;
  isUnlocked: boolean;
  progressPercent: number;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  task_earning: 'task_earning',
  commission: 'commission',
  streak_bonus: 'streak_bonus',
  daily_yield: 'daily_yield',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  completed: 'completed',
  rejected: 'rejected',
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  description: string;
  status: TransactionStatus;
  createdAt: string;
}

export interface UserProfileUpdate {
  fullName?: string;
  email?: string;
}

export interface DeleteAccountInput {
  confirmationText: string;
}

export type ListTransactionsParams = {
type?: ListTransactionsType;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType];


export const ListTransactionsType = {
  all: 'all',
  deposits: 'deposits',
  withdrawals: 'withdrawals',
  task_earnings: 'task_earnings',
  commissions: 'commissions',
} as const;