# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BirrStream API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /packages/{id}/purchase:
    post:
      operationId: purchasePackage
      tags: [packages]
      summary: Purchase a VIP package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseResult"
        "400":
          description: Insufficient balance or validation error
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform-wide stats (admin only)
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Not an admin
  /admin/deposits/pending:
    get:
      operationId: listPendingDeposits
      tags: [admin]
      summary: List all pending deposits (admin only)
      responses:
        "200":
          description: Pending deposits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminDeposit"
        "403":
          description: Not an admin
  /admin/deposits/{id}/approve:
    post:
      operationId: approveDeposit
      tags: [admin]
      summary: Approve a pending deposit and credit the user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Not an admin
        "404":
          description: Deposit not found
  /admin/deposits/{id}/reject:
    post:
      operationId: rejectDeposit
      tags: [admin]
      summary: Reject a pending deposit (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Not an admin
        "404":
          description: Deposit not found
  /admin/withdrawals/pending:
    get:
      operationId: listPendingWithdrawals
      tags: [admin]
      summary: List all pending withdrawals (admin only)
      responses:
        "200":
          description: Pending withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminWithdrawal"
        "403":
          description: Not an admin
  /admin/withdrawals/{id}/approve:
    post:
      operationId: approveWithdrawal
      tags: [admin]
      summary: Approve a pending withdrawal (funds already deducted) (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Not an admin
        "404":
          description: Withdrawal not found
  /admin/withdrawals/{id}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [admin]
      summary: Reject a pending withdrawal and refund the user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal rejected and refunded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Not an admin
        "404":
          description: Withdrawal not found
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        mainBalance:
          type: number
        totalYield:
          type: number
        totalDeposited:
          type: number
        totalWithdrawn:
          type: number
        referralCode:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - fullName
        - username
        - email
        - mainBalance
        - totalYield
        - totalDeposited
        - totalWithdrawn
        - referralCode
        - isAdmin
        - createdAt
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    RegisterInput:
      type: object
      properties:
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
        referralCode:
          type: ["string", "null"]
      required:
        - fullName
        - username
        - email
        - password
        - confirmPassword
    LoginInput:
      type: object
      properties:
        usernameOrEmail:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
      required:
        - usernameOrEmail
        - password
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
      required:
        - email
...
    Package:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cost:
          type: number
        dailyReturn:
          type: number
        totalYield:
          type: number
        durationDays:
          type: integer
        isLocked:
          type: boolean
        tier:
          type: string
          enum: [vip1, vip2, vip3, vip4, vip5, elite, apex, titan, alpha]
      required:
        - id
        - name
        - cost
        - dailyReturn
        - totalYield
        - durationDays
        - isLocked
        - tier
    PurchaseResult:
      type: object
      properties:
        success:
          type: boolean
        package:
          $ref: "#/components/schemas/Package"
        newBalance:
          type: number
        message:
          type: string
        shortfallAmount:
          type: ["number", "null"]
      required:
        - success
        - message
...
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalDeposited:
          type: number
        totalWithdrawn:
          type: number
        pendingDepositsCount:
          type: integer
        pendingDepositsAmount:
          type: number
        pendingWithdrawalsCount:
          type: integer
        pendingWithdrawalsAmount:
          type: number
        totalActivePackages:
          type: integer
      required:
        - totalUsers
        - totalDeposited
        - totalWithdrawn
        - pendingDepositsCount
        - pendingDepositsAmount
        - pendingWithdrawalsCount
        - pendingWithdrawalsAmount
        - totalActivePackages
    AdminDeposit:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        fullName:
          type: string
        amount:
          type: number
        senderName:
          type: string
        receiptUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
      required:
        - id
        - userId
        - username
        - fullName
        - amount
        - senderName
        - status
        - createdAt
    AdminWithdrawal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        fullName:
          type: string
        amount:
          type: number
        bankName:
          type: string
        accountName:
          type: string
        walletId:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
      required:
        - id
        - userId
        - username
        - fullName
        - amount
        - bankName
        - accountName
        - walletId
        - status
        - createdAt
...