# lib/api-zod/src/generated/types/adminWithdrawal.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BirrStream API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminWithdrawalStatus } from './adminWithdrawalStatus';

export interface AdminWithdrawal {
  id: number;
  userId: number;
  username: string;
  fullName: string;
  amount: number;
  bankName: string;
  accountName: string;
  walletId: string;
  status: AdminWithdrawalStatus;
  createdAt: string;
}