# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BirrStream API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  fullName: string;
  username: string;
  email: string;
  mainBalance: number;
  totalYield: number;
  totalDeposited: number;
  totalWithdrawn: number;
  referralCode: string;
  isAdmin: boolean;
  createdAt: string;
}